requirements.txt
blinker==1.9.0
click==8.1.8
Flask==3.1.0
importlib_metadata==8.5.0
itsdangerous==2.2.0
Jinja2==3.1.5
joblib==1.4.2
MarkupSafe==3.0.2
numpy==2.0.2
pillow==11.1.0
scikit-learn==1.6.0
scipy==1.13.1
threadpoolctl==3.5.0
Werkzeug==3.1.3
zipp==3.21.0
